Created a math header file
In lieu of copying the lambda for checking primality whenever there is a
new problem that deals with primes, let's use a header file to keep
track of all of the functions that may be recycled in the future.

This has been implemented in math.hpp, and the functionality ported
throughout the already solved problems.