Adding (very simple) matrix functionality to math
Before we start tackling graph theory problems, it would be nice to have
a library to handle matricies. This has only been lightly tested, and
implements the features that I'm sure I'll need. Other functionality
will be added as needed.

I have also included a basic "test" file that was used for testing this.
This does no negative testing (most of that was done accidentally during
the process of testing), and shouldn't reflect any real-world testing
practices.